# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Cambase.io is a project by Evercam.io in order to make it easier for\
    \ software developers to have a reliable, up to date source of model hardware\
    \ information available via a public API."
  version: "1.0"
  title: "Cambase.io"
  x-logo:
    url: "https://api.apis.guru/v2/cache/logo/https_twitter.com_evrcm_profile_image.jpeg"
  x-origin:
  - converter:
      url: "https://github.com/lucybot/api-spec-converter"
      version: "2.6.2"
    format: "swagger"
    url: "http://api.cambase.io/api-docs/api-docs.json"
    version: "1.2"
  x-preferred: true
  x-providerName: "cambase.io"
host: "api.cambase.io"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /api/v1/vendors/{id}.json:
    get:
      tags:
      - "vendors"
      summary: "Fetches a single Vendor"
      operationId: "Api::V1::Vendors#show"
      parameters:
      - name: "id"
        in: "path"
        description: "Vendor ID"
        required: true
        type: "string"
      - name: "order"
        in: "query"
        description: "Sort order"
        required: false
        type: "string"
      responses:
        "200":
          description: "No response was specified"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "406":
          description: "Not Acceptable"
      x-swagger-router-controller: "Vendors"
    put:
      tags:
      - "vendors"
      summary: "Updates an existing Vendor"
      operationId: "apiV1VendorsId_jsonPUT"
      parameters:
      - name: "id"
        in: "path"
        description: "Vendor ID"
        required: true
        type: "string"
      - name: "vendor[name]"
        in: "formData"
        description: "Name"
        required: false
        type: "string"
      - name: "vendor[info]"
        in: "formData"
        description: "Info."
        required: false
        type: "string"
      - name: "vendor[url]"
        in: "formData"
        description: "Website"
        required: false
        type: "string"
      - name: "vendor[mac]"
        in: "formData"
        description: "MAC"
        required: false
        type: "string"
      responses:
        "200":
          description: "No response was specified"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "406":
          description: "Not Acceptable"
      x-swagger-router-controller: "Vendors"
    patch:
      tags:
      - "vendors"
      summary: "Updates an existing Vendor"
      operationId: "apiV1VendorsId_jsonPATCH"
      parameters:
      - name: "id"
        in: "path"
        description: "Vendor ID"
        required: true
        type: "string"
      - name: "vendor[name]"
        in: "formData"
        description: "Name"
        required: false
        type: "string"
      - name: "vendor[info]"
        in: "formData"
        description: "Info."
        required: false
        type: "string"
      - name: "vendor[url]"
        in: "formData"
        description: "Website"
        required: false
        type: "string"
      - name: "vendor[mac]"
        in: "formData"
        description: "MAC"
        required: false
        type: "string"
      responses:
        "200":
          description: "No response was specified"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "406":
          description: "Not Acceptable"
      x-swagger-router-controller: "Vendors"